requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Pillow==9.1.0
# psycopg2-binary==2.9.3
drf-yasg==1.20.0
django-filter